# pos-carts/src/main/resources/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Cart
  license:
    name: MIT
paths:
  /cart/add:
    post:
      summary: Add item to cart
      operationId: add item
      requestBody:
        description: add item to cart
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemFields'
        required: true
      responses:
        '200':
          description: Add succeed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Items"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/del:
    post:
      summary: Del the existence or decrease the quantity of certain item in the cart
      operationId: del item
      requestBody:
        description: del item in cart
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemFields'
        required: true
      responses:
        '200':
          description: Del succeed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Items'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /cart/order:
    post:
      summary: Add item to cart
      operationId: makeAnOrder
      requestBody:
        description: add item to cart
        content:
          application/json:
            schema:
              type: object
              properties:
                cart:
                  $ref: '#/components/schemas/Items'
                usrId:
                  type: string
        required: true
      responses:
        '200':
          description: Add succeed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Items"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    Item:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
        quantity:
          type: integer
          format: int32
    Items:
      type: array
      items:
        $ref: "#/components/schemas/Item"
    ItemFields:
      title: item field
      type: object
      description: field for cart CRUD.
      allOf:
        - $ref: '#/components/schemas/Item'
        - type: object
          properties:
            cart:
              title: Current Cart
              type: array
              items:
                $ref: "#/components/schemas/Item"
      required:
        - productId
        - quantity
        - cart
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string